pyproject.toml
[project]
name = "classifier-assignment"
version = "0.1.0"
description = "Classifier"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
   "fastapi>=0.115.12",
   "python-multipart>=0.0.20",
   "uvicorn>=0.34.2",
   "PyPDF2>=3.0.0", 
   "pdf2image>=1.17.0",
   "PyMuPDF>=1.23.1",
   "pytesseract>=0.3.10",
   "pillow>=10.0.0",
   "ultralytics>=8.2.28",
   "dill>=0.3.8",
]

[dependency-groups]
dev = [
    "ruff>=0.11.5",
]
